fix 'settings invalid' error when loading config with 'first byte is size byte' enabled
It is expected that 'frameSizeChanged' signal is emitted twice. First when
'frame size' field is set, second time when 'fixed size' button is unchecked.
But this doesn't happen because 'fixed size' radio button is already unchecked.
'toggled' signal is only emitted if state of the button actually changes.
We manually emit the 'frameSizeChanged' signal to make sure final state of the
config is reflected on the 'FramedReader' instance.